requirements.txt
instabot==0.4.5
python-dotenv==0.10.1